pyproject.toml
[project]
name = "pana-201-p002"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.3",
    "rich>=14.1.0",
    "tavily-python>=0.7.10",
]
